+++ src/steps/all.ts
@@ -1,5 +1,6 @@
 import './step-id'
 import './input-field-dialog'
 import './event-call-history'
+import './action-history'
 import './entity-selection/entity-selection'
 import './entity-selection/teleport'