+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Fix remote sc.CombatProxyEntity not sharing animation state
 - Fix ig.ENTITY.ObjectLayerView being invisible
+- Fix no sound playing on client when the map instance is hidden
 
 ## [0.5.8] 2025-11-13